requirements.txt
tensorflow>=2.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
matplotlib>=3.5.0
